Implement navigation system
To render the navigation bar, pass a list of `url_for`[1]-compatible
endpoint names in the `nav` property of your desired template, like so:

    return render_template("file.html", nav=["index.root", "index.files"])

and include the _nav.html file in the template itself:

    {%- include '_nav.html' %}

Each endpoint must have a registered title. In order to register a title,
decorate the endpoint function with @nav.register_title("Page title").

[1]: https://flask.palletsprojects.com/en/2.0.x/api/#flask.url_for